testing out a dedicated function library (eatlib)
